# docs/openapi.yaml
---
openapi: 3.1.0
tags:
- name: Contact Groups
  description: Operations for managing groups
- name: Contacts
  description: Operations for managing contacts
components:
  schemas:
    ContactGroupRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          pattern: \S
          maxLength: 100
        description:
          type: string
          maxLength: 500
    ContactGroupResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
    ContactRequest:
      type: object
      required:
      - firstName
      - lastName
      - email
      - owner
      properties:
        firstName:
          type: string
          pattern: \S
          maxLength: 100
        lastName:
          type: string
          pattern: \S
          maxLength: 100
        email:
          type: string
          pattern: \S
          maxLength: 255
        phone:
          type: string
          maxLength: 50
        owner:
          type: string
          pattern: \S
          maxLength: 100
        groupId:
          type: integer
          format: int64
    ContactResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        owner:
          type: string
        groupId:
          type: integer
          format: int64
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
    LocalDateTime:
      type: string
      format: date-time
      examples:
      - 2022-03-10T12:15:50
    PagedResponseContactGroupResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContactGroupResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
    PagedResponseContactResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContactResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
paths:
  /api/v1/contacts:
    get:
      summary: Get all contacts
      description: Returns a paginated and filtered list of contacts
      tags:
      - Contacts
      parameters:
      - name: owner
        in: query
        schema:
          type: string
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          minimum: 0
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          maximum: 100
          minimum: 1
          default: 10
      responses:
        "200":
          description: Contacts retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseContactResponse"
        "400":
          description: Invalid pagination parameters
    post:
      summary: Create a contact
      description: Creates a new contact
      tags:
      - Contacts
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactRequest"
        required: true
      responses:
        "201":
          description: Contact created successfully
          headers:
            Location:
              description: URL of the created contact
              schema:
                type: string
                format: uri
        "400":
          description: Invalid contact data
        "404":
          description: Contact group not found
        "409":
          description: Contact email already exists
  /api/v1/contacts/{id}:
    put:
      summary: Update a contact
      description: Updates an existing contact
      tags:
      - Contacts
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactRequest"
        required: true
      responses:
        "200":
          description: Contact updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "400":
          description: Invalid contact data
        "404":
          description: Contact or contact group not found
        "409":
          description: Contact email already exists
    get:
      summary: Get contact by ID
      description: Returns a contact by its ID
      tags:
      - Contacts
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Contact retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "404":
          description: Contact not found
    delete:
      summary: Delete a contact
      description: Deletes a contact by its ID
      tags:
      - Contacts
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Contact deleted successfully
        "404":
          description: Contact not found
  /api/v1/groups:
    get:
      summary: Get all groups
      description: Returns a paginated and filtered list of groups
      tags:
      - Contact Groups
      parameters:
      - name: name
        in: query
        schema:
          type: string
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          minimum: 0
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          maximum: 100
          minimum: 1
          default: 10
      responses:
        "200":
          description: Groups retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseContactGroupResponse"
        "400":
          description: Invalid pagination parameters
    post:
      summary: Create a group
      description: Creates a new group
      tags:
      - Contact Groups
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactGroupRequest"
        required: true
      responses:
        "201":
          description: Group created successfully
          headers:
            Location:
              description: URL of the created group
              schema:
                type: string
                format: uri
        "400":
          description: Invalid group data
        "409":
          description: Group name already exists
  /api/v1/groups/{id}:
    put:
      summary: Update a group
      description: Updates an existing group
      tags:
      - Contact Groups
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactGroupRequest"
        required: true
      responses:
        "200":
          description: Group updated successfully
        "400":
          description: Invalid group data
        "404":
          description: Group not found
        "409":
          description: Group name already exists
    get:
      summary: Get group by ID
      description: Returns a contact by its ID
      tags:
      - Contact Groups
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Group retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactGroupResponse"
        "404":
          description: Group not found
    delete:
      summary: Delete a group
      description: Deletes a group by its ID
      tags:
      - Contact Groups
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Group deleted successfully
        "404":
          description: Group not found
  /api/v1/groups/{id}/contacts:
    get:
      summary: Get contacts by group
      description: Returns contacts belonging to a specific group
      tags:
      - Contact Groups
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          minimum: 0
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
          default: 10
      responses:
        "200":
          description: Group contacts retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseContactResponse"
        "400":
          description: Invalid pagination parameters
        "404":
          description: Group not found
info:
  title: contacts-api API
  version: 1.0.0-SNAPSHOT